Initial shift to double.

+++ EniseySolution/Enisey/edge.h
@@ -12,8 +12,8 @@ public:
   virtual void set_gas_out(const Gas* gas) = 0;
   virtual const Gas& gas_in() = 0;
   virtual const Gas& gas_out() = 0;
-  virtual float q() = 0;
-  virtual float dq_dp_in() = 0;
-  virtual float dq_dp_out() = 0;
-  virtual float IsReverse() = 0;
+  virtual double q() = 0;
+  virtual double dq_dp_in() = 0;
+  virtual double dq_dp_out() = 0;
+  virtual double IsReverse() = 0;
 };
