+++ EniseySolution/Enisey/edge_model_pipe_sequential.cpp
@@ -10,16 +10,16 @@ void EdgeModelPipeSequential::set_gas_in(const Gas* gas) {
 void EdgeModelPipeSequential::set_gas_out(const Gas* gas) {
   pointer_to_model_->set_gas_out(gas);
 }
-float EdgeModelPipeSequential::q() {
+double EdgeModelPipeSequential::q() {
   return pointer_to_model_->q();
 }
-float EdgeModelPipeSequential::dq_dp_in() {
+double EdgeModelPipeSequential::dq_dp_in() {
   return pointer_to_model_->dq_dp_in();
 }
-float EdgeModelPipeSequential::dq_dp_out() {
+double EdgeModelPipeSequential::dq_dp_out() {
   return pointer_to_model_->dq_dp_out();
 }
-float EdgeModelPipeSequential::IsReverse() {
+double EdgeModelPipeSequential::IsReverse() {
   return  pointer_to_model_->IsReverse();
 }
 const Gas& EdgeModelPipeSequential::gas_in() {
