+++ EniseySolution/Enisey/gas.h
@@ -3,16 +3,16 @@
 
 struct GasWorkParameters {
   GasWorkParameters() : p(0), t(0), q(0) { };
-  float p;	// Давление, [МПа]
-  float t;	// Тепереатура, [К]
-  float q;	// Количество газа, [млн м3/сут]
+  double p;	// Давление, [МПа]
+  double t;	// Тепереатура, [К]
+  double q;	// Количество газа, [млн м3/сут]
 }; 
 
 struct GasCompositionReduced {
   GasCompositionReduced() : density_std_cond(0), n2(0), co2(0) { };
-  float density_std_cond;		// Плотность при стд. усл-ях [кг/м3]
-  float n2;					// Доля азота [б.р.]
-  float co2;					// Доля углекислого газа [б.р.]
+  double density_std_cond;		// Плотность при стд. усл-ях [кг/м3]
+  double n2;					// Доля азота [б.р.]
+  double co2;					// Доля углекислого газа [б.р.]
 };
 
 struct Gas {
@@ -23,9 +23,9 @@ struct Gas {
   Расход может быть отрицательным, если труба реверсивная. Раз потоки всё же
   смешиваются, берём абсолютное значение расходов. */
   void Mix(const Gas &mix) {  
-    float q1 = abs(work_parameters.q);
-    float q2 = abs(mix.work_parameters.q);
-    float sum_q = q1 + q2;
+    double q1 = abs(work_parameters.q);
+    double q2 = abs(mix.work_parameters.q);
+    double sum_q = q1 + q2;
     composition.density_std_cond = // ro1 = (q1*ro1 + q2*ro2) / (q1 + q2)
         (q1 * composition.density_std_cond + 
         q2 * mix.composition.density_std_cond ) / sum_q;
