+++ EniseySolution/Enisey/gas_count_functions.cuh
@@ -5,25 +5,25 @@
 
 // Базовые свойства газа
 // Найти псеводокритическую температуру [K]
-inline float FindTPseudoCritical(float den_sc, float co2, float n2)
+inline double FindTPseudoCritical(double den_sc, double co2, double n2)
 {
 	return 88.25 * (0.9915 + 1.759 * den_sc - co2 - 1.681*n2); // [Кельвины]	
 };
 
 // Найти псевдокритическое давление [МПа]
-inline float FindPPseudoCritical(float den_sc, float co2, float n2)
+inline double FindPPseudoCritical(double den_sc, double co2, double n2)
 {
 	return 2.9585 * (1.608 - 0.05994*den_sc - co2 - 0.392*n2); // [МегаПаскали]
 };
 
 // Фактор сжимаемости при с.у. [б.р.]
-inline float FindZStandartConditions(float den_sc, float co2, float n2)
+inline double FindZStandartConditions(double den_sc, double co2, double n2)
 {
 	return 1 - pow(0.0741*den_sc-0.006-0.063*n2-0.0575*co2, 2); // [б.р.]
 };
 
 // Газовая постоянная при с.у. [Джоуль / (Моль*Кельвин)]
-inline float FindRStandartConditions(float den_sc)
+inline double FindRStandartConditions(double den_sc)
 {
 	// kAirDensityStandartConditions = 1.2046
 	return 286.89 * 1.2046 / den_sc; // [Дж/(моль*К)]
@@ -31,76 +31,76 @@ inline float FindRStandartConditions(float den_sc)
 
 // Далее - функции вычисления параметров при рабочих давлении и тем-ре
 // Вычислить приведённые давление и температуру
-inline float FindPReduced(float p_work, float p_pseudo_critical)
+inline double FindPReduced(double p_work, double p_pseudo_critical)
 {
 	return p_work / p_pseudo_critical; // [б.р.]
 };
 	
-inline float FindTReduced(float t_work, float t_pseudo_critical)
+inline double FindTReduced(double t_work, double t_pseudo_critical)
 {
 	return t_work / t_pseudo_critical; // [б.р.]
 };
 
 // Теплоёмкость при р.у. [Дж/(кг*К)]
-inline float FindC(float t_reduced, float p_reduced, float r_standart_conditions)
+inline double FindC(double t_reduced, double p_reduced, double r_standart_conditions)
 {
-	float E0 = 4.437 - 1.015*t_reduced + 0.591*t_reduced*t_reduced;
-	float E1 = 3.29  - 11.37/t_reduced + 10.9 /(t_reduced*t_reduced);
-	float E2 = 3.23  - 16.27/t_reduced + 25.48/(t_reduced*t_reduced) 
+	double E0 = 4.437 - 1.015*t_reduced + 0.591*t_reduced*t_reduced;
+	double E1 = 3.29  - 11.37/t_reduced + 10.9 /(t_reduced*t_reduced);
+	double E2 = 3.23  - 16.27/t_reduced + 25.48/(t_reduced*t_reduced) 
 		- 11.81/(t_reduced*t_reduced*t_reduced);
-	float E3 = -0.214 + 0.908/t_reduced -0.967/(t_reduced*t_reduced);
+	double E3 = -0.214 + 0.908/t_reduced -0.967/(t_reduced*t_reduced);
 
 	return r_standart_conditions * (E0 + E1*p_reduced + E2*(p_reduced*p_reduced) + 
 		E3*(p_reduced*p_reduced*p_reduced) ); // [Дж/(кг*К)]
 };
 
 // Коэффициент Джоуля-Томпсона при рабочих условиях
-inline float FindDi(float p_reduced, float t_reduced)
+inline double FindDi(double p_reduced, double t_reduced)
 {
-	float H0 = 24.94 - 20.3*t_reduced  + 4.57  *(t_reduced*t_reduced);
-	float H1 = 5.66  - 19.92/t_reduced + 16.89 /(t_reduced*t_reduced);
-	float H2 = -4.11 + 14.68/t_reduced - 13.39 /(t_reduced*t_reduced);
-	float H3 = 0.568 - 2.0/t_reduced   + 1.79  /(t_reduced*t_reduced);
+	double H0 = 24.94 - 20.3*t_reduced  + 4.57  *(t_reduced*t_reduced);
+	double H1 = 5.66  - 19.92/t_reduced + 16.89 /(t_reduced*t_reduced);
+	double H2 = -4.11 + 14.68/t_reduced - 13.39 /(t_reduced*t_reduced);
+	double H3 = 0.568 - 2.0/t_reduced   + 1.79  /(t_reduced*t_reduced);
 
-	float di_ = H0 + H1*p_reduced + H2*(p_reduced*p_reduced) + 
+	double di_ = H0 + H1*p_reduced + H2*(p_reduced*p_reduced) + 
 		H3*(p_reduced*p_reduced*p_reduced); 
 	di_ *= 0.000001; // ToDo: разбраться с этим множителем
 	return di_;
 };
 
 // Динамическая вязкость при рабочих условиях
-inline float FindMju(float p_reduced, float t_reduced)
+inline double FindMju(double p_reduced, double t_reduced)
 {
-	float Mju0 = (1.81 + 5.95*t_reduced); 
-	float B1 = -0.67 + 2.36/t_reduced  - 1.93  / (t_reduced*t_reduced);
-	float B2 = 0.8   - 2.89/t_reduced  + 2.65  / (t_reduced*t_reduced);
-	float B3 = -0.1  + 0.354/t_reduced - 0.314 / (t_reduced*t_reduced);
+	double Mju0 = (1.81 + 5.95*t_reduced); 
+	double B1 = -0.67 + 2.36/t_reduced  - 1.93  / (t_reduced*t_reduced);
+	double B2 = 0.8   - 2.89/t_reduced  + 2.65  / (t_reduced*t_reduced);
+	double B3 = -0.1  + 0.354/t_reduced - 0.314 / (t_reduced*t_reduced);
 
-	float mju_ = Mju0 * (1 + B1*p_reduced + B2*(p_reduced*p_reduced) + 
+	double mju_ = Mju0 * (1 + B1*p_reduced + B2*(p_reduced*p_reduced) + 
 		B3*(p_reduced*p_reduced*p_reduced));
 	mju_ /= 1000000; // ToDo: разобраться с этим множителем
 	return mju_;
 };
 
 // Коэффициент сжимаемости при рабочих условиях [б.р.]
-inline float FindZ(float p_reduced, float t_reduced)
+inline double FindZ(double p_reduced, double t_reduced)
 {
-	float A1 = -0.39 + 2.03/t_reduced - 3.16/(t_reduced*t_reduced) 
+	double A1 = -0.39 + 2.03/t_reduced - 3.16/(t_reduced*t_reduced) 
 		+ 1.09/(t_reduced*t_reduced*t_reduced);
-	float A2 = 0.0423 - 0.1812/t_reduced + 0.2124/(t_reduced*t_reduced);
+	double A2 = 0.0423 - 0.1812/t_reduced + 0.2124/(t_reduced*t_reduced);
 	return 1 + A1*p_reduced + A2*(p_reduced*p_reduced);
 };
 
 // Плотность при рабочих условиях [кг/м3]
-inline float FindRo(float den_sc, float p_work, float t_work, float z)
+inline double FindRo(double den_sc, double p_work, double t_work, double z)
 {
-	//static const float kTemperatureStandartConditions = 293.15; // [K] 
-	//static const float  kPressureStandartConditions = 0.101325; // [MПа] 
+	//static const double kTemperatureStandartConditions = 293.15; // [K] 
+	//static const double  kPressureStandartConditions = 0.101325; // [MПа] 
 	return (den_sc * 293.15 * p_work) / (t_work*0.101325*z);
 };	
 
 // Число Рейнольдса
-inline float FindRe(float q, float den_sc, float mju, float d_inner)
+inline double FindRe(double q, double den_sc, double mju, double d_inner)
 {
 	// kPi = 3.14159265358979323846264338327950288419716939937510
 	return (4.0/3.14159265358979323846264338327950288419716939937510) * (q*den_sc) / (mju*(d_inner/1000));
@@ -108,10 +108,10 @@ inline float FindRe(float q, float den_sc, float mju, float d_inner)
 
 // Коэффициент гидравлического сопротивления (число Лямбда) 
 // Требует, чтобы число Рейнольдса было рассчитано!
-inline float FindLambda(float re, float d_inner, float roughness_coefficient, float hydraulic_efficiency_coefficient)
+inline double FindLambda(double re, double d_inner, double roughness_coefficient, double hydraulic_efficiency_coefficient)
 {
-	float osn = 158.0/re + (2*roughness_coefficient/d_inner);
-	float lambda = 0.067 * exp(0.2 * log(osn)); 
+	double osn = 158.0/re + (2*roughness_coefficient/d_inner);
+	double lambda = 0.067 * exp(0.2 * log(osn)); 
 	// ToDo: возможно стоит хранить сразу квадрат к-та гидравлич. эф-ти
 	lambda /= pow(hydraulic_efficiency_coefficient, 2); 
 	return lambda;
